Force db_auto_create to False in bin/glance-manage
When interacting with the db migration tool, bin/glance-manage,
we cannot depend on the behavior of auto-creating database tables.
This plays havoc with the migrations that also add those tables. It
is better practice to prevent anything automatic from happening
while running db_sync and to explicitly depend on the migration
scripts.

Fixes bug 1042925

Change-Id: Id93283905256247bdd2663b799d41a7925804542